+++ packages/components/env.d.ts
@@ -1,4 +1,4 @@
-/// <reference types="@sv/elements" />
+/// <reference types="@svp/elements" />
 
 declare namespace JSX {
   // TODO: Polyfill type for JSX.Element