+++ packages/components/src/vue/Combobox.tsx
@@ -1,8 +1,8 @@
 /* @jsxImportSource vue */
 
-import "@sv/elements/dropdown";
-import "@sv/elements/toggle";
-import "@sv/elements/expandable";
+import "@svp/elements/dropdown";
+import "@svp/elements/toggle";
+import "@svp/elements/expandable";
 
 import { Button } from "./Button.jsx";
 import { Icon } from "./Icon.jsx";