+++ packages/components/src/vue/Popover.tsx
@@ -1,6 +1,6 @@
 /* @jsxImportSource vue */
 
-import "@sv/elements/popover";
+import "@svp/elements/popover";
 import { Button } from "./Button";
 
 export function Popover(props: { label?: string }, { slots }) {