+++ packages/elements/packages/dropdown/src/index.ts
@@ -1,4 +1,4 @@
-import "@sv/elements/expandable";
+import "@svp/elements/expandable";
 
 export { Dropdown } from "./components/dropdown/Dropdown";
 export { OptionElement } from "./components/dropdown/Option";