+++ packages/components/src/vue/Accordion.tsx
@@ -1,6 +1,6 @@
 /* @jsxImportSource vue */
 
-import "@svp/elements/expandable";
+import "@sv/elements/expandable";
 import { defineComponent } from "vue";
 import { Icon } from "./Icon.jsx";