+++ packages/elements/packages/popover/src/Popover.spec.ts
@@ -3,7 +3,7 @@ import { test, expect } from "bun:test";
 const NODE_NAME = "a-popover";
 
 test("import element", async () => {
-  const { Popover } = await import("@svp/elements/popover");
+  const { Popover } = await import("@sv/elements/popover");
   expect(Popover).toBeDefined();
 
   // is defined in custom element registry
@@ -20,7 +20,7 @@ test("import element", async () => {
 });
 
 test("initial state", async () => {
-  const { Popover } = await import("@svp/elements/popover");
+  const { Popover } = await import("@sv/elements/popover");
   const ele = new Popover();
 
   expect(ele.opened).toBe(false);