+++ docs/.config/vite.config.ts
@@ -4,7 +4,7 @@ export default defineConfig({
   resolve: {
     alias: {
       "package:": "/src",
-      "@components": "@sv/components",
+      "@components": "@atrium-ui/components",
     },
   },
   server: {