+++ packages/scroll-lock/test/index.spec.ts
@@ -2,7 +2,7 @@ import { describe, expect, it } from "bun:test";
 
 describe("scroll-lock import", () => {
   it("import module", async () => {
-    const out = await import(require.resolve("@sv/scroll-lock"));
+    const out = await import(require.resolve("@atrium-ui/scroll-lock"));
     expect(out.ScrollLock).toBeDefined();
   });
 });