+++ packages/html-to-custom-tag-replacement/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @sv/html-to-custom-tag-replacement
 
+## 0.1.2 (2024-11-14)
+
+### Fixes
+
+- Make accordion content inert if closed
+
 ## 0.1.1 (2024-04-13)
 
 ### Features